Move sync progress functionality to a new module
We may need to implement multiple ways of calculating sync progress.
Having it in its own module makes it easier to grasp.
