+++ lib/jormungandr/src/Cardano/Wallet/Jormungandr.hs
@@ -119,6 +119,8 @@ import Cardano.Wallet.Primitive.AddressDiscovery.Random
     ( RndState )
 import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
     ( SeqState )
+import Cardano.Wallet.Primitive.SyncProgress
+    ( SyncTolerance )
 import Cardano.Wallet.Primitive.Types
     ( Address
     , Block
@@ -127,7 +129,6 @@ import Cardano.Wallet.Primitive.Types
     , GenesisParameters (..)
     , NetworkParameters (..)
     , ProtocolParameters (..)
-    , SyncTolerance
     , WalletId
     )
 import Cardano.Wallet.Registry
